Use google java style with checkstyle plugin
In order to work with the latest version of the google_checks.xml file,
checkstyle version 6.4.1 was required, however the lastest maven
checkstyle plugin uses version 6.1.1. In order to upgrade Checkstyle at
runtime, the latest version is pulled in as a dependency. In order to
pull in dependencies, the maven-dependency-plugin had to be added.

The maven-jxr-plugin was also added to enable cross-referencing of
source code in the checkstyle report.