fixed bug with enable/disable menuitems by cherry-picking from next story's branch. story:40
CHERRY-PICKED COMMIT FROM feature/skillsRegister

Refactor menu bar

Previously the state (i.e. enabled/disabled) of menu bar buttons
was being initialized in the FXML file. This was confusing. So I created
a new initial function in in the menu bar controller which sets the
initial state of the buttons.

Conflicts:
	src/main/resources/menubar.fxml