Large changes to the way tab content is handled.
Peoples list view is not populated upon a project selection change, (previously it was populated when clicking on the People tab - this was very bad since you could switch between tabs and the content would repeatedly be reloaded, even when it didn't change. Same for skills). Also added skills validation on creating a new skill (Previously you could create duplicated skills, only validation was performed on editing - this is now fixed.
story:41