Refactoring
Project model now also has transient ObservableList for People/Teams/Skills. This directly connects with the ArrayList of People/Teams/Skills so any update in the ObservableList is updated in the ArrayList. The result is cleaner and easier code.

Operations happen in the Command itself now, not overwritten by an anonymous Command declaration.

For now saving upon creating of objects is removed, as this is being changed from the current implementation.
story:131