Give sequence a key
Begin report with key for sequence. For example if printing a project
report, the sequence will have the key `Projects:`. This is necessary to
make the yaml parsable by common parsers such as yamljs.
Also changed comment from "LEVEL" to "<CLASS> REPORT", as I think this
makes more sense.

story[171]