Refactor dialog title setting and field population
Previously the stage title was being set in the populateFields method of
each form controller. We already had the infrastructure in place to
dynamically set the stages title but it was slightly broken. Fixed the
method in main controller for setting stage titles and removed title
setting from the population fields method. Also small refactorings to
make code simpler.
story:204