Reduce coupling between StoryView and Model
Refactored StoryFormController so that it is less coupled with the
model. Previously it was setting its textFields by accessing the model
directly. This commit makes it so, the ViewModel accesses the model and
the View binds its text fields to the ViewModels properties.

However, I have left the projectTextField property as is because it was
not easy to refactor without breaking things. It should be changed in
the future however.

I also bound the editable property of the creatorTextField to a new
boolean creatorEditable property of the ViewModel.

story:204