Move all validation to the ViewModel
Previously only the validation predicate was declared in the ViewModel.
This moves things closer to the MVVM ideal by placing all of the
validation logic (and even error messages) in the ViewModel and
binding the ControlsFX validation visualisers to the ViewModel validators
boolean and error message properties.

This change also required modifying the unit tests (for the better).
The viewModel now needs to be altered when testing different cases
(for example the shortNameProperty needs to be set for each short name
to be tested). Also, the boolean values being test are those of the ViewModel's
validators. The good thing about this is that the unit tests are now
more closely aligned to the View and the properties it depends on.

Even the okButton (enable/disable) validation has been moved to the
ViewModel. The ViewModel contains a composite validator which combines
all of the other validators. The okButtons disable property is bound to
the inverse of the boolean value of this composite validator.

story:204