Split MenuBarController into View/ViewModel
Unfortunately the tab list's toggle group's selectedToggleProperty() is
read only. This means that the selected toggle cannot not be changed
using bindings (not easily atleast). For this reason I had to add a
listener on the viewModel's selected tab property to the View.
This is not ideal, but it works.

As far as unit testing is concerned, most of the actions are hooks into
mainController and sideBarController. These controllers have not been
moved to MVVM so they still run in a JavaFX thread and cannot
easily be unit tested. Therefore I haven't added any unit tests.

story:279