Sprint View/ViewModel: Re-engineering
This commits introduces two significant changes:

1. The most significant change is wrapping the model using the mvvmfx
[ModelWrapper class](https://github.com/sialcasa/mvvmFX/wiki/ModelWrapper).
This allows the removal of significant ammounts of boilerplate.
The model wrapper also provides an isDifferent flag which indicates
whether or not the ViewModels properties differ from that
Unfortunately, the wrapper does not currently support observable lists
very well, and so these have to be checked seperately. There is an
[issue](https://github.com/sialcasa/mvvmFX/issues/221) on GitHub for
adding observable list support to ModelWrapper and the mvvmfx are very
welcoming of pull requests ;)

2. Update sprint view to use new string converter methods.
The new string converter methods allow an ObjectProperty<Organisation>
to passed through as a parameter. As a result, the organisation used for
validation can be dynamically changed after a string binding is created.
This allows the for the removal of significant ammounts of boilerplate
from the ViewViewModel. The View can now bind the textProperty of it's
TextFields and Labels directly to the ViewModel's ObjectProperties,
passing through the ViewModel's organisationProperty as a parameter to
the StringConverter.

Unit tests are also added.

story:288