SprintViewModel: Prevent null pointer exception
When creating a new Sprint and leaving the optional description field
blank, the form would throw a null pointer exception when opened for
editing. This is because the model wrapper was setting the value to null
rather than empty string.

This is fixed by calling reset() (and reseting the model wrappers values
to the sane defaults) before calling commit().

story:288