Story details pane re-engineered
Implemented the model wrapper in the storyViewModel.
Updated storyFormView to work with the model wrapper.
Added storyFormViewModel.
Refactored the handling of ok and cancel button actions and implemented the strategy design pattern.
Then updated  the dialog method in MainController to use the FluentViewLoader.
Moved FXML into a parallel directory structure.

 story:288