Add bound property support
Bound property support makes it possible to add listeners to model
classes which fire whenever any of the observable properties of that
class change. This class was found on GitHub:
https://github.com/BenjaminGale/Bound-Property-Support

Bound property support is added to the Item class.
The BoundPropertySupport variable is made protected so that children can
access it to add property change listeners.