GoatLabel: new interface and utility methods
The main change to GoatLabel is the addition of two event handlers:
onAction and onCancel. The onAction method is invoked whenever the editable
node is fired (e.g. when the done button is clicked or enter pressed).
The onCancel method is invoked whenenver an in progress edit is
cancelled (e.g. a user clicks outside of the editable node).
These methods can be set using the GoatLabels setOnAction() and
setOnCancel() methods.

This makes it possible to specify what actions a ViewModel/Controller should
take whenever an edit is made or cancelled. For example, when an edit is
cancelled, we can call the ViewModel's reload method to ensure that the
label shows the original value. When an edit is made we can call the
UndoManager from the ViewModel. This means that only the ViewModel/Controllers
need to be aware of UndoManager.

For convenience, a utility function initGoatLabel was added to FxUtils.
This function makes it possible to initialise GoatLabels in the View
using one line. Currently, initGoatLabel is only implemented for TextFields
and TextAreas, however, it should be fairly easy to implement the method for
DatePicker, ComboBox, etc.

The SkillDetailsView/ViewModel has been updated to show how this new
interface works in practice.

story:355