initGoatLabel extensions
Add support for using ObjectProperties with GoatLabelTextFields and
GoatLabelTextAreas. Also add support for GoatLabelComboBoxes.
If the initGoatLabel utility method ever gets superseeded, it will be by
something which should be should be super easy to plug in, so I'm going
to go ahead and keep extending this, even if we end up using something
different.

story:355