Add ListSupport to BoundPropertySupport
By default, BoundPropertySupport only adds a changed listener to a
ListProperty. This means that there is no list change listener on the
ListProperties observable list, so the no changeEvent is fired when the
contents of the list changes.

This commit adds a ListChangeListener which is attached to any
ListProperties which are added. However, BoundPropertySupport still only
binds to a ChangeListener, so we have to somehow convert the
ListChangeListener to a ChangeListener which has observable, oldValue,
and newValue properties.

Keeping track of these things is too much work, so the values are
fudged. At this point, we only use BoundPropertySupport to know that
something changed. We don't care what changed or what its old and new
values are. Therefore, this fudging of values is okay for now.

story:355