BacklogVieWModel: Potential command improvements
Instead of iterating through each field with if-else statements, we
can just get a list of changed fields from the ModelWrapper and then
create edit commands using a lambda expression. Unfortunately this does not
work so well for fields that have back references, so these are still
handled seperately.

story:355