Add Effort to model.
Effort is not searchable because we currently do not have any place
to "take" a user when they click on the an efforts search result.

Both logTime and Task are immutable (i.e. they only provide a getter
method) meaning that they can not be changed after an Effort object
is created.

story:342