Fix: "New Project" hangs application
Fixes the issue that occurs when creating a new project when internet
is not enabled. See #75. The sidebar controller was using the controlsfx
library to get it's fontawesome icons. This library attempts to connect
to CDN and thus hangs on the lab machine when the internet is not enabled.
Now that this issue is fixed, the application will segfault when creating
a new project if the internet is not enabled.

story:285