PersistenceManagerTest: add explanatory comment
Adds a comment explaining why we do not unit test loading a Sprint1 file.
Thought that the test was missing a statement, added it, and then found
out why we don't properly unit test it. Thought I should add a comment
so that others don't wasted time with this in the future.

story:345