TaskCard: Fix blocked property highlighting issue
Fixes issue where changing the blocked property in the expanded task card
was not changing the blocked property visualisation (i.e. 'glow') on the
scrumboard. This was due to an edit command not being created because
the ViewModel was using it's own modelWrapper rather than that of its
parent class.

Also adds unit tests for this case.

story:285